Preparing to porting to OSX and iOS. MACOSX renamed to OSX.
Fix include typos.

Show message char set fix.

Code formatting.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@522 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 2b30341a227a4c7d5b94ecddadd7392224a46b76

+++ Source/Debugger.cpp
@@ -561,7 +561,7 @@ StackInfo* GetStackInfo( const void* caller )
     str.reserve( 16384 );
 
     char buf[ 64 ];
-    sprintf( buf, "\t0x%08X\n\n", (size_t) caller );
+    sprintf( buf, "\t0x%08X\n\n", (uint) caller );
     str += buf;
 
     StackInfo* si = new StackInfo();
@@ -690,7 +690,7 @@ string Debugger::GetTraceMemory()
     for( auto it = blocks_chunks.begin(), end = blocks_chunks.end(); it != end; ++it )
     {
         StackInfo* si = *it;
-        sprintf( buf, "%06u) Size %-10u Chunks %-10u Heap %08X\n", cur++, si->Size, si->Chunks, si->Heap );
+        sprintf( buf, "%06u) Size %-10u Chunks %-10u Heap %08X\n", cur++, si->Size, si->Chunks, (uint) si->Heap );
         str += buf;
         str += si->Name;
     }
